moved shared test definitions to com.google.idea.testing.headless
