+++ clwb/tests/integrationtests/com/google/idea/blaze/clwb/VirtualIncludesTest.java
@@ -4,8 +4,8 @@ import static com.google.common.truth.Truth.assertThat;
 import static com.google.idea.blaze.clwb.base.Assertions.assertContainsHeader;
 
 import com.google.idea.blaze.base.bazel.BazelVersion;
-import com.google.idea.blaze.clwb.base.ProjectViewBuilder;
-import com.google.idea.blaze.clwb.base.ClwbIntegrationTestCase;
+import com.google.idea.blaze.clwb.base.ClwbHeadlessTestCase;
+import com.google.idea.testing.headless.ProjectViewBuilder;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.jetbrains.cidr.lang.workspace.OCCompilerSettings;
 import org.jetbrains.annotations.Nullable;
@@ -14,7 +14,7 @@ import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
 @RunWith(JUnit4.class)
-public class VirtualIncludesTest extends ClwbIntegrationTestCase {
+public class VirtualIncludesTest extends ClwbHeadlessTestCase {
 
   @Test
   public void testClwb() {
