+++ clwb/tests/integrationtests/com/google/idea/blaze/clwb/base/ClwbHeadlessTestCase.java
@@ -1,0 +1,32 @@
+package com.google.idea.blaze.clwb.base;
+
+import static com.google.common.truth.Truth.assertThat;
+
+import com.google.idea.testing.headless.HeadlessTestCase;
+import com.jetbrains.cidr.lang.CLanguageKind;
+import com.jetbrains.cidr.lang.workspace.OCCompilerSettings;
+import com.jetbrains.cidr.lang.workspace.OCResolveConfiguration;
+import com.jetbrains.cidr.lang.workspace.OCWorkspace;
+
+public class ClwbHeadlessTestCase extends HeadlessTestCase {
+
+  protected OCWorkspace getWorkspace() {
+    return OCWorkspace.getInstance(myProject);
+  }
+
+  protected OCResolveConfiguration findFileResolveConfiguration(String relativePath) {
+    final var file = findProjectFile(relativePath);
+
+    final var configurations = getWorkspace().getConfigurationsForFile(file);
+    assertThat(configurations).hasSize(1);
+
+    return configurations.get(0);
+  }
+
+  protected OCCompilerSettings findFileCompilerSettings(String relativePath) {
+    final var file = findProjectFile(relativePath);
+    final var resolveConfiguration = findFileResolveConfiguration(relativePath);
+
+    return resolveConfiguration.getCompilerSettings(CLanguageKind.CPP, file);
+  }
+}
