+++ testing/src/com/google/idea/testing/headless/InMemoryPreferencesFactory.java
@@ -1,4 +1,4 @@
-package com.google.idea.blaze.clwb.base;
+package com.google.idea.testing.headless;
 
 import java.util.HashMap;
 import java.util.Map;
