+++ testing/src/com/google/idea/testing/headless/OSRule.java
@@ -1,4 +1,4 @@
-package com.google.idea.blaze.clwb.base;
+package com.google.idea.testing.headless;
 
 import java.util.Arrays;
 import org.junit.rules.TestRule;
