+++ testing/src/com/google/idea/testing/headless/Statements.java
@@ -1,4 +1,4 @@
-package com.google.idea.blaze.clwb.base;
+package com.google.idea.testing.headless;
 
 import org.junit.runners.model.Statement;
